Prune elixir phoenix example
