+++ example-servers/elixir/phoenix/assets/js/app.js
@@ -22,6 +22,7 @@ import { Socket } from "phoenix";
 import { LiveSocket } from "phoenix_live_view";
 import topbar from "../vendor/topbar";
 import "deep-chat";
+import "../css/app.css";
 
 let csrfToken = document
   .querySelector("meta[name='csrf-token']")
