+++ example-servers/elixir/phoenix/config/config.exs
@@ -7,50 +7,28 @@
 # General application configuration
 import Config
 
-config :deep_chat_phoenix_lw,
-  ecto_repos: [DeepChatPhoenixLw.Repo],
+config :deep_chat_phoenixlv_ex,
   generators: [timestamp_type: :utc_datetime]
 
 # Configures the endpoint
-config :deep_chat_phoenix_lw, DeepChatPhoenixLwWeb.Endpoint,
+config :deep_chat_phoenixlv_ex, DeepChatPhoenixlvExWeb.Endpoint,
   url: [host: "localhost"],
   adapter: Bandit.PhoenixAdapter,
   render_errors: [
-    formats: [html: DeepChatPhoenixLwWeb.ErrorHTML, json: DeepChatPhoenixLwWeb.ErrorJSON],
+    formats: [html: DeepChatPhoenixlvExWeb.ErrorHTML, json: DeepChatPhoenixlvExWeb.ErrorJSON],
     layout: false
   ],
-  pubsub_server: DeepChatPhoenixLw.PubSub,
-  live_view: [signing_salt: "Wzb85cN/"]
-
-# Configures the mailer
-#
-# By default it uses the "Local" adapter which stores the emails
-# locally. You can see the emails in your browser, at "/dev/mailbox".
-#
-# For production it's recommended to configure a different adapter
-# at the `config/runtime.exs`.
-config :deep_chat_phoenix_lw, DeepChatPhoenixLw.Mailer, adapter: Swoosh.Adapters.Local
+  pubsub_server: DeepChatPhoenixlvEx.PubSub,
+  live_view: [signing_salt: "mBRrfbDv"]
 
 # Configure esbuild (the version is required)
 config :esbuild,
   version: "0.17.11",
-  deep_chat_phoenix_lw: [
+  deep_chat_phoenixlv_ex: [
     args:
       ~w(js/app.js --bundle --target=es2017 --outdir=../priv/static/assets --external:/fonts/* --external:/images/*),
     cd: Path.expand("../assets", __DIR__),
     env: %{"NODE_PATH" => Path.expand("../deps", __DIR__)}
-  ]
-
-# Configure tailwind (the version is required)
-config :tailwind,
-  version: "3.4.3",
-  deep_chat_phoenix_lw: [
-    args: ~w(
-      --config=tailwind.config.js
-      --input=css/app.css
-      --output=../priv/static/assets/app.css
-    ),
-    cd: Path.expand("../assets", __DIR__)
   ]
 
 # Configures Elixir's Logger
