+++ example-servers/elixir/phoenix/config/prod.exs
@@ -5,14 +5,8 @@ import Config
 # manifest is generated by the `mix assets.deploy` task,
 # which you should run after static files are built and
 # before starting your production server.
-config :deep_chat_phoenix_lw, DeepChatPhoenixLwWeb.Endpoint,
+config :deep_chat_phoenixlv_ex, DeepChatPhoenixlvExWeb.Endpoint,
   cache_static_manifest: "priv/static/cache_manifest.json"
-
-# Configures Swoosh API Client
-config :swoosh, api_client: Swoosh.ApiClient.Finch, finch_name: DeepChatPhoenixLw.Finch
-
-# Disable Swoosh Local Memory Storage
-config :swoosh, local: false
 
 # Do not print debug messages in production
 config :logger, level: :info
