+++ example-servers/elixir/phoenix/config/runtime.exs
@@ -12,30 +12,15 @@ import Config
 # If you use `mix release`, you need to explicitly enable the server
 # by passing the PHX_SERVER=true when you start it:
 #
-#     PHX_SERVER=true bin/deep_chat_phoenix_lw start
+#     PHX_SERVER=true bin/deep_chat_phoenixlv_ex start
 #
 # Alternatively, you can use `mix phx.gen.release` to generate a `bin/server`
 # script that automatically sets the env var above.
 if System.get_env("PHX_SERVER") do
-  config :deep_chat_phoenix_lw, DeepChatPhoenixLwWeb.Endpoint, server: true
+  config :deep_chat_phoenixlv_ex, DeepChatPhoenixlvExWeb.Endpoint, server: true
 end
 
 if config_env() == :prod do
-  database_url =
-    System.get_env("DATABASE_URL") ||
-      raise """
-      environment variable DATABASE_URL is missing.
-      For example: ecto://USER:PASS@HOST/DATABASE
-      """
-
-  maybe_ipv6 = if System.get_env("ECTO_IPV6") in ~w(true 1), do: [:inet6], else: []
-
-  config :deep_chat_phoenix_lw, DeepChatPhoenixLw.Repo,
-    # ssl: true,
-    url: database_url,
-    pool_size: String.to_integer(System.get_env("POOL_SIZE") || "10"),
-    socket_options: maybe_ipv6
-
   # The secret key base is used to sign/encrypt cookies and other secrets.
   # A default value is used in config/dev.exs and config/test.exs but you
   # want to use a different value for prod and you most likely don't want
@@ -51,9 +36,9 @@ if config_env() == :prod do
   host = System.get_env("PHX_HOST") || "example.com"
   port = String.to_integer(System.get_env("PORT") || "4000")
 
-  config :deep_chat_phoenix_lw, :dns_cluster_query, System.get_env("DNS_CLUSTER_QUERY")
+  config :deep_chat_phoenixlv_ex, :dns_cluster_query, System.get_env("DNS_CLUSTER_QUERY")
 
-  config :deep_chat_phoenix_lw, DeepChatPhoenixLwWeb.Endpoint,
+  config :deep_chat_phoenixlv_ex, DeepChatPhoenixlvExWeb.Endpoint,
     url: [host: host, port: 443, scheme: "https"],
     http: [
       # Enable IPv6 and bind on all interfaces.
@@ -70,7 +55,7 @@ if config_env() == :prod do
   # To get SSL working, you will need to add the `https` key
   # to your endpoint configuration:
   #
-  #     config :deep_chat_phoenix_lw, DeepChatPhoenixLwWeb.Endpoint,
+  #     config :deep_chat_phoenixlv_ex, DeepChatPhoenixlvExWeb.Endpoint,
   #       https: [
   #         ...,
   #         port: 443,
@@ -92,26 +77,8 @@ if config_env() == :prod do
   # We also recommend setting `force_ssl` in your config/prod.exs,
   # ensuring no data is ever sent via http, always redirecting to https:
   #
-  #     config :deep_chat_phoenix_lw, DeepChatPhoenixLwWeb.Endpoint,
+  #     config :deep_chat_phoenixlv_ex, DeepChatPhoenixlvExWeb.Endpoint,
   #       force_ssl: [hsts: true]
   #
   # Check `Plug.SSL` for all available options in `force_ssl`.
-
-  # ## Configuring the mailer
-  #
-  # In production you need to configure the mailer to use a different adapter.
-  # Also, you may need to configure the Swoosh API client of your choice if you
-  # are not using SMTP. Here is an example of the configuration:
-  #
-  #     config :deep_chat_phoenix_lw, DeepChatPhoenixLw.Mailer,
-  #       adapter: Swoosh.Adapters.Mailgun,
-  #       api_key: System.get_env("MAILGUN_API_KEY"),
-  #       domain: System.get_env("MAILGUN_DOMAIN")
-  #
-  # For this example you need include a HTTP client required by Swoosh API client.
-  # Swoosh supports Hackney and Finch out of the box:
-  #
-  #     config :swoosh, :api_client, Swoosh.ApiClient.Hackney
-  #
-  # See https://hexdocs.pm/swoosh/Swoosh.html#module-installation for details.
 end
