--- example-servers/elixir/phoenix/lib/deep_chat_phoenix_lw.ex
@@ -1,9 +1,0 @@
-defmodule DeepChatPhoenixLw do
-  @moduledoc """
-  DeepChatPhoenixLw keeps the contexts that define your domain
-  and business logic.
-
-  Contexts are also responsible for managing your data, regardless
-  if it comes from the database, an external API or others.
-  """
-end
