--- example-servers/elixir/phoenix/lib/deep_chat_phoenix_lw/mailer.ex
@@ -1,3 +1,0 @@
-defmodule DeepChatPhoenixLw.Mailer do
-  use Swoosh.Mailer, otp_app: :deep_chat_phoenix_lw
-end
