--- example-servers/elixir/phoenix/lib/deep_chat_phoenix_lw/repo.ex
@@ -1,5 +1,0 @@
-defmodule DeepChatPhoenixLw.Repo do
-  use Ecto.Repo,
-    otp_app: :deep_chat_phoenix_lw,
-    adapter: Ecto.Adapters.Postgres
-end
