--- example-servers/elixir/phoenix/lib/deep_chat_phoenix_lw_web/components/layouts.ex
@@ -1,14 +1,0 @@
-defmodule DeepChatPhoenixLwWeb.Layouts do
-  @moduledoc """
-  This module holds different layouts used by your application.
-
-  See the `layouts` directory for all templates available.
-  The "root" layout is a skeleton rendered as part of the
-  application router. The "app" layout is set as the default
-  layout on both `use DeepChatPhoenixLwWeb, :controller` and
-  `use DeepChatPhoenixLwWeb, :live_view`.
-  """
-  use DeepChatPhoenixLwWeb, :html
-
-  embed_templates "layouts/*"
-end
