--- example-servers/elixir/phoenix/lib/deep_chat_phoenix_lw_web/controllers/page_controller.ex
@@ -1,9 +1,0 @@
-defmodule DeepChatPhoenixLwWeb.PageController do
-  use DeepChatPhoenixLwWeb, :controller
-
-  def home(conn, _params) do
-    # The home page is often custom made,
-    # so skip the default app layout.
-    render(conn, :home, layout: false)
-  end
-end
