--- example-servers/elixir/phoenix/lib/deep_chat_phoenix_lw_web/controllers/page_html.ex
@@ -1,10 +1,0 @@
-defmodule DeepChatPhoenixLwWeb.PageHTML do
-  @moduledoc """
-  This module contains pages rendered by PageController.
-
-  See the `page_html` directory for all templates available.
-  """
-  use DeepChatPhoenixLwWeb, :html
-
-  embed_templates "page_html/*"
-end
