+++ example-servers/elixir/phoenix/lib/deep_chat_phoenixlv_ex.ex
@@ -1,0 +1,9 @@
+defmodule DeepChatPhoenixlvEx do
+  @moduledoc """
+  DeepChatPhoenixlvEx keeps the contexts that define your domain
+  and business logic.
+
+  Contexts are also responsible for managing your data, regardless
+  if it comes from the database, an external API or others.
+  """
+end
