+++ example-servers/elixir/phoenix/lib/deep_chat_phoenixlv_ex/application.ex
@@ -1,0 +1,33 @@
+defmodule DeepChatPhoenixlvEx.Application do
+  # See https://hexdocs.pm/elixir/Application.html
+  # for more information on OTP Applications
+  @moduledoc false
+
+  use Application
+
+  @impl true
+  def start(_type, _args) do
+    children = [
+      DeepChatPhoenixlvExWeb.Telemetry,
+      {DNSCluster, query: Application.get_env(:deep_chat_phoenixlv_ex, :dns_cluster_query) || :ignore},
+      {Phoenix.PubSub, name: DeepChatPhoenixlvEx.PubSub},
+      # Start a worker by calling: DeepChatPhoenixlvEx.Worker.start_link(arg)
+      # {DeepChatPhoenixlvEx.Worker, arg},
+      # Start to serve requests, typically the last entry
+      DeepChatPhoenixlvExWeb.Endpoint
+    ]
+
+    # See https://hexdocs.pm/elixir/Supervisor.html
+    # for other strategies and supported options
+    opts = [strategy: :one_for_one, name: DeepChatPhoenixlvEx.Supervisor]
+    Supervisor.start_link(children, opts)
+  end
+
+  # Tell Phoenix to update the endpoint configuration
+  # whenever the application is updated.
+  @impl true
+  def config_change(changed, _new, removed) do
+    DeepChatPhoenixlvExWeb.Endpoint.config_change(changed, removed)
+    :ok
+  end
+end
