+++ example-servers/elixir/phoenix/lib/deep_chat_phoenixlv_ex_web/components/layouts.ex
@@ -1,0 +1,14 @@
+defmodule DeepChatPhoenixlvExWeb.Layouts do
+  @moduledoc """
+  This module holds different layouts used by your application.
+
+  See the `layouts` directory for all templates available.
+  The "root" layout is a skeleton rendered as part of the
+  application router. The "app" layout is set as the default
+  layout on both `use DeepChatPhoenixlvExWeb, :controller` and
+  `use DeepChatPhoenixlvExWeb, :live_view`.
+  """
+  use DeepChatPhoenixlvExWeb, :html
+
+  embed_templates "layouts/*"
+end
