+++ example-servers/elixir/phoenix/lib/deep_chat_phoenixlv_ex_web/components/layouts/app.html.heex
@@ -1,0 +1,5 @@
+<main>
+  <div class="main-container">
+    {@inner_content}
+  </div>
+</main>
