+++ example-servers/elixir/phoenix/lib/deep_chat_phoenixlv_ex_web/endpoint.ex
@@ -1,0 +1,43 @@
+defmodule DeepChatPhoenixlvExWeb.Endpoint do
+  use Phoenix.Endpoint, otp_app: :deep_chat_phoenixlv_ex
+
+  # The session will be stored in the cookie and signed,
+  # this means its contents can be read but not tampered with.
+  # Set :encryption_salt if you would also like to encrypt it.
+  @session_options [
+    store: :cookie,
+    key: "_deep_chat_phoenixlv_ex_key",
+    signing_salt: "7wryr4Ad",
+    same_site: "Lax"
+  ]
+
+  socket "/live", Phoenix.LiveView.Socket,
+    websocket: [connect_info: [session: @session_options]],
+    longpoll: [connect_info: [session: @session_options]]
+
+  # Serve at "/" the static files from "priv/static" directory.
+  #
+  # You should set gzip to true if you are running phx.digest
+  # when deploying your static files in production.
+  plug Plug.Static,
+    at: "/",
+    from: :deep_chat_phoenixlv_ex,
+    gzip: false,
+    only: DeepChatPhoenixlvExWeb.static_paths()
+
+  # Code reloading can be explicitly enabled under the
+  # :code_reloader configuration of your endpoint.
+  if code_reloading? do
+    socket "/phoenix/live_reload/socket", Phoenix.LiveReloader.Socket
+    plug Phoenix.LiveReloader
+    plug Phoenix.CodeReloader
+  end
+
+  plug Plug.RequestId
+  plug Plug.Telemetry, event_prefix: [:phoenix, :endpoint]
+
+  plug Plug.MethodOverride
+  plug Plug.Head
+  plug Plug.Session, @session_options
+  plug DeepChatPhoenixlvExWeb.Router
+end
