+++ example-servers/elixir/phoenix/mix.exs
@@ -1,9 +1,9 @@
-defmodule DeepChatPhoenixLw.MixProject do
+defmodule DeepChatPhoenixlvEx.MixProject do
   use Mix.Project
 
   def project do
     [
-      app: :deep_chat_phoenix_lw,
+      app: :deep_chat_phoenixlv_ex,
       version: "0.1.0",
       elixir: "~> 1.14",
       elixirc_paths: elixirc_paths(Mix.env()),
@@ -18,7 +18,7 @@ defmodule DeepChatPhoenixLw.MixProject do
   # Type `mix help compile.app` for more information.
   def application do
     [
-      mod: {DeepChatPhoenixLw.Application, []},
+      mod: {DeepChatPhoenixlvEx.Application, []},
       extra_applications: [:logger, :runtime_tools]
     ]
   end
@@ -33,29 +33,14 @@ defmodule DeepChatPhoenixLw.MixProject do
   defp deps do
     [
       {:phoenix, "~> 1.7.14"},
-      {:phoenix_ecto, "~> 4.5"},
-      {:ecto_sql, "~> 3.10"},
-      {:postgrex, ">= 0.0.0"},
       {:phoenix_html, "~> 4.1"},
       {:phoenix_live_reload, "~> 1.2", only: :dev},
       # TODO bump on release to {:phoenix_live_view, "~> 1.0.0"},
       {:phoenix_live_view, "~> 1.0.0-rc.1", override: true},
       {:floki, ">= 0.30.0", only: :test},
-      {:phoenix_live_dashboard, "~> 0.8.3"},
       {:esbuild, "~> 0.8", runtime: Mix.env() == :dev},
-      {:tailwind, "~> 0.2", runtime: Mix.env() == :dev},
-      {:heroicons,
-       github: "tailwindlabs/heroicons",
-       tag: "v2.1.1",
-       sparse: "optimized",
-       app: false,
-       compile: false,
-       depth: 1},
-      {:swoosh, "~> 1.5"},
-      {:finch, "~> 0.13"},
       {:telemetry_metrics, "~> 1.0"},
       {:telemetry_poller, "~> 1.0"},
-      {:gettext, "~> 0.20"},
       {:jason, "~> 1.2"},
       {:dns_cluster, "~> 0.1.1"},
       {:bandit, "~> 1.5"}
@@ -70,15 +55,11 @@ defmodule DeepChatPhoenixLw.MixProject do
   # See the documentation for `Mix` for more info on aliases.
   defp aliases do
     [
-      setup: ["deps.get", "ecto.setup", "assets.setup", "assets.build"],
-      "ecto.setup": ["ecto.create", "ecto.migrate", "run priv/repo/seeds.exs"],
-      "ecto.reset": ["ecto.drop", "ecto.setup"],
-      test: ["ecto.create --quiet", "ecto.migrate --quiet", "test"],
-      "assets.setup": ["tailwind.install --if-missing", "esbuild.install --if-missing"],
-      "assets.build": ["tailwind deep_chat_phoenix_lw", "esbuild deep_chat_phoenix_lw"],
+      setup: ["deps.get", "assets.setup", "assets.build"],
+      "assets.setup": ["esbuild.install --if-missing"],
+      "assets.build": ["esbuild deep_chat_phoenixlv_ex"],
       "assets.deploy": [
-        "tailwind deep_chat_phoenix_lw --minify",
-        "esbuild deep_chat_phoenix_lw --minify",
+        "esbuild deep_chat_phoenixlv_ex --minify",
         "phx.digest"
       ]
     ]
