--- example-servers/elixir/phoenix/priv/repo/seeds.exs
@@ -1,11 +1,0 @@
-# Script for populating the database. You can run it as:
-#
-#     mix run priv/repo/seeds.exs
-#
-# Inside the script, you can read and write to any of your
-# repositories directly:
-#
-#     DeepChatPhoenixLw.Repo.insert!(%DeepChatPhoenixLw.SomeSchema{})
-#
-# We recommend using the bang functions (`insert!`, `update!`
-# and so on) as they will fail if something goes wrong.
