--- example-servers/elixir/phoenix/test/deep_chat_phoenix_lw_web/controllers/error_html_test.exs
@@ -1,14 +1,0 @@
-defmodule DeepChatPhoenixLwWeb.ErrorHTMLTest do
-  use DeepChatPhoenixLwWeb.ConnCase, async: true
-
-  # Bring render_to_string/4 for testing custom views
-  import Phoenix.Template
-
-  test "renders 404.html" do
-    assert render_to_string(DeepChatPhoenixLwWeb.ErrorHTML, "404", "html", []) == "Not Found"
-  end
-
-  test "renders 500.html" do
-    assert render_to_string(DeepChatPhoenixLwWeb.ErrorHTML, "500", "html", []) == "Internal Server Error"
-  end
-end
