--- example-servers/elixir/phoenix/test/deep_chat_phoenix_lw_web/controllers/page_controller_test.exs
@@ -1,8 +1,0 @@
-defmodule DeepChatPhoenixLwWeb.PageControllerTest do
-  use DeepChatPhoenixLwWeb.ConnCase
-
-  test "GET /", %{conn: conn} do
-    conn = get(conn, ~p"/")
-    assert html_response(conn, 200) =~ "Peace of mind from prototype to production"
-  end
-end
